[virtio-serial] Fix 64 bit compilation

+++ virtio-serial/VIOSerialCoreQueue.c
@@ -329,7 +329,7 @@ NTSTATUS VSCSendCopyBuffer(PVIOSERIAL_PORT pPort,
 
 NTSTATUS VSCRecieveCopyBuffer(PVIOSERIAL_PORT pPort,
 							  WDFMEMORY * buffer,
-							  unsigned int * pSize,
+							  size_t * pSize,
 							  PKSPIN_LOCK pLock)
 {
 	NTSTATUS status = STATUS_SUCCESS;
